+++ jingle_sleigh/Cargo.toml
@@ -24,7 +24,6 @@ cxx = "1.0.122"
 serde = { version = "1.0.203", features = ["derive"] }
 serde-xml-rs = "0.6.0"
 thiserror = { version = "1.0.61", features = [] }
-elf = { version = "0.7.4", optional = true }
 object = { version = "0.36.0", optional = true }
 libz-sys = { version = "1.1.18", default-features = false, features = ["libc"] }
 tracing = "0.1.40"
@@ -33,8 +32,7 @@ tracing = "0.1.40"
 cxx-build = "1.0.122"
 
 [features]
-elf = ["dep:elf"]
 gimli = ["dep:object"]
-default = ["elf", "gimli"]
+default = ["gimli"]
