+++ jingle_sleigh/src/ffi/context_ffi.rs
@@ -1,9 +1,11 @@
 use std::pin::Pin;
 use crate::ffi::context_ffi::bridge::makeContext;
 use bridge::ContextFFI;
-use cxx::{Exception, UniquePtr};
+use cxx::{Exception, ExternType, UniquePtr};
 use std::sync::Mutex;
 use crate::context::image::ImageProvider;
+use crate::ffi::instruction::bridge::VarnodeInfoFFI;
+use crate::VarNode;
 
 type ContextGeneratorFp = fn(&str) -> Result<UniquePtr<ContextFFI>, Exception>;
 
@@ -57,4 +59,24 @@ pub(crate) mod bridge {
 
 pub(crate) struct ImageFFI<'a> {
     pub(crate) provider: Pin<Box<dyn ImageProvider + 'a>>,
+}
+
+impl<'a> ImageFFI<'a> {
+    pub(crate) fn new<T: ImageProvider + 'a>(provider: T) -> Self {
+        Self {
+            provider: Box::pin(provider),
+        }
+    }
+    pub(crate) fn load(&self, vn: &VarnodeInfoFFI, out: &mut [u8]) -> usize {
+        self.provider.load(&VarNode::from(vn), out)
+    }
+
+    pub(crate) fn has_range(&self, vn: &VarNode) -> bool {
+        self.provider.has_full_range(vn)
+    }
+}
+
+unsafe impl<'a> ExternType for ImageFFI<'a> {
+    type Id = cxx::type_id!("ImageFFI");
+    type Kind = cxx::kind::Opaque;
 }
